data_processing/requirements.txt
librosa>=0.10.0
soundfile>=0.12.0
numpy>=1.24.0

openai-whisper>=20230314

torch>=2.0.0
torchaudio>=2.0.0

pandas>=2.0.0
tqdm>=4.65.0

pathlib2>=2.3.7

audeer
audonnx
